# Optimize advanced detector by 34% in perf