This PR improves the advanced detector algorithm running time by 34% from 3000us to 2400us. This was measured on iptsd-perf using Linux perf.

I ran the profiler on iptsd-perf and found it strange that 25% of the time was spent on std::hypotf. I took a look at the code only to find that function was only called on constants. I tried adding `inline` to `get_cost`, but the compiler still wouldn't inline it. Therefore, I forced the compiler to inline things using template parameters and `constexpr`. This improved the performance by 20%.

`std::reference_wrapper` is used in the class desugared from the lambda by request. It initially caused a small degradation, until I did the following. In the disassembly, I see a lot of calls related to std::vector, so I determined that bounds checks were taking a lot of time. I applied the `common::unchecked` pattern to this lambda, and the nearby ones. This improved the performance by a further 17%.

These measurements were done on `iptsd-perf` on a separate machine. With the actual Surface Pro 7 device, and the actual `iptsd` daemon, it was around 50%+ CPU usage before, and around 45% after. I would have investigated that discrepancy if I had a working `perf`. It's the same 34% in perf, but a 10% in `iptsd`. The actual numbers follow:

<details>
<summary>On-device numbers</summary>

```
# 0 commits
[03:19:06.555] [info] Vendor:       045E
[03:19:06.555] [info] Product:      099F
[03:19:06.555] [info] Buffer Size:  7487
[03:19:06.555] [info] Metadata:
[03:19:06.555] [info] rows=44, columns=64
[03:19:06.555] [info] width=25978, height=17319
[03:19:06.555] [info] transform=[-412.3492,0,25978,0,-402.76746,17319]
[03:19:06.555] [info] unknown=1, [178,182,180,1,178,182,180,1,90,171,100,20,172,177,175,2]
[03:19:14.614] [info] Ran 2010 times
[03:19:14.614] [info] Total: 8042632μs
[03:19:14.614] [info] Mean: 4001.31μs
[03:19:14.614] [info] Standard Deviation: 1399.72μs
[03:19:14.614] [info] Minimum: 181.138μs
[03:19:14.614] [info] Maximum: 7771.524μs
# 1 commit
[03:17:18.905] [info] Vendor:       045E
[03:17:18.905] [info] Product:      099F
[03:17:18.905] [info] Buffer Size:  7487
[03:17:18.905] [info] Metadata:
[03:17:18.905] [info] rows=44, columns=64
[03:17:18.905] [info] width=25978, height=17319
[03:17:18.905] [info] transform=[-412.3492,0,25978,0,-402.76746,17319]
[03:17:18.905] [info] unknown=1, [178,182,180,1,178,182,180,1,90,171,100,20,172,177,175,2]
[03:17:25.786] [info] Ran 2010 times
[03:17:25.786] [info] Total: 6860737μs
[03:17:25.786] [info] Mean: 3413.30μs
[03:17:25.786] [info] Standard Deviation: 1174.28μs
[03:17:25.786] [info] Minimum: 180.452μs
[03:17:25.786] [info] Maximum: 6039.696μs
# commits
[03:16:14.245] [info] Vendor:       045E
[03:16:14.245] [info] Product:      099F
[03:16:14.245] [info] Buffer Size:  7487
[03:16:14.245] [info] Metadata:
[03:16:14.245] [info] rows=44, columns=64
[03:16:14.245] [info] width=25978, height=17319
[03:16:14.245] [info] transform=[-412.3492,0,25978,0,-402.76746,17319]
[03:16:14.245] [info] unknown=1, [178,182,180,1,178,182,180,1,90,171,100,20,172,177,175,2]
[03:16:19.668] [info] Ran 2010 times
[03:16:19.668] [info] Total: 5399899μs
[03:16:19.668] [info] Mean: 2686.52μs
[03:16:19.668] [info] Standard Deviation: 906.13μs
[03:16:19.668] [info] Minimum: 177.923μs
[03:16:19.668] [info] Maximum: 4784.426μs
```
</details>

# Before

```
/home/home/CLionProjects/iptsd/build/src/debug/iptsd-perf ./iptsdump.dat
[08:03:23.027] [info] Vendor:       045E
[08:03:23.027] [info] Product:      099F
[08:03:23.027] [info] Buffer Size:  7487
[08:03:23.027] [info] Metadata:
[08:03:23.027] [info] rows=44, columns=64
[08:03:23.027] [info] width=25978, height=17319
[08:03:23.027] [info] transform=[-412.3492,0,25978,0,-402.76746,17319]
[08:03:23.027] [info] unknown=1, [178,182,180,1,178,182,180,1,90,171,100,20,172,177,175,2]
[08:03:29.048] [info] Ran 2010 times
[08:03:29.048] [info] Total: 6011079μs
[08:03:29.048] [info] Mean: 2990.59μs
[08:03:29.048] [info] Standard Deviation: 1045.43μs
[08:03:29.048] [info] Minimum: 147.510μs
[08:03:29.048] [info] Maximum: 5385.695μs
```

![image](https://user-images.githubusercontent.com/25646384/220932567-e2182518-97d5-4ba3-8be3-1cdad03396b8.png)

# After

```
[22:20:12.187] [info] Vendor:       045E
[22:20:12.187] [info] Product:      099F
[22:20:12.187] [info] Buffer Size:  7487
[22:20:12.187] [info] Metadata:
[22:20:12.187] [info] rows=44, columns=64
[22:20:12.187] [info] width=25978, height=17319
[22:20:12.187] [info] transform=[-412.3492,0,25978,0,-402.76746,17319]
[22:20:12.187] [info] unknown=1, [178,182,180,1,178,182,180,1,90,171,100,20,172,177,175,2]
[22:20:16.228] [info] Ran 2010 times
[22:20:16.228] [info] Total: 4031356μs
[22:20:16.228] [info] Mean: 2005.65μs
[22:20:16.228] [info] Standard Deviation: 674.57μs
[22:20:16.228] [info] Minimum: 144.730μs
[22:20:16.228] [info] Maximum: 3357.086μs
```

![image](https://user-images.githubusercontent.com/25646384/221084140-3017ed53-dffb-4143-be3b-09323a030b6d.png)