@qzed 2023-02-23
Ah, that's a good catch. Somewhat disappointing that it didn't inline/pre-compute it, especially since it was a lambda...

One note though: I'm not sure how much time we should spend on optimizing the "advanced" processor. I think it's unlikely to get it down to a reasonable performance level