@StollD 2023-02-23
> Somewhat disappointing that it didn't inline/pre-compute it, especially since it was a lambda...

Maybe the reason for that is that the `weighted_distance_transform` function is not marked as inline? It seperates the lambda from the (inline) function that calls it.