@qzed 2023-02-23
I don't think that should be an issue. Although I'm not entirely sure about what the standard prescribes, the lambda should have a unique type, so `weighted_distance_transform()` should get specifically instantiated for that lambda. The problem is that the compiler might recognize the lambda as non-trivial function, so it doesn't inline it, which might be the reason it misses out on the optimizations. So maybe one could try to force-inline the lambda, but I'm not sure if that is actually beneficial in the end...