@danielzgtg 2023-02-24
I just switched to a new branch and made a few test edits. It is still 3000us after putting `[[gnu::always_inline]] inline ` in front of every single function in `distance_transform.hpp`. Putting that only in front of `weighted_distance_transform()` gives the same result.

I was thinking it's because hypotf isn't constexpr, but turns out the compiler [can still inline that](https://godbolt.org/#g:!((g:!((g:!((h:codeEditor,i:(filename:'1',fontScale:14,fontUsePx:'0',j:1,lang:c%2B%2B,selection:(endColumn:2,endLineNumber:10,positionColumn:2,positionLineNumber:10,selectionStartColumn:2,selectionStartLineNumber:10,startColumn:2,startLineNumber:10),source:'%23include+%3Ccmath%3E%0A%23include+%3Ccstdio%3E%0A%0Afloat+test(int+x,+int+y)+%7B%0A++++return+hypotf(x,+y)%3B%0A%7D%0A%0Aint+main()+%7B%0A++++printf(%22%25f%22,+test(1,+2))%3B%0A%7D'),l:'5',n:'0',o:'C%2B%2B+source+%231',t:'0')),k:50,l:'4',n:'0',o:'',s:0,t:'0'),(g:!((h:compiler,i:(compiler:g122,deviceViewOpen:'1',filters:(b:'0',binary:'1',binaryObject:'1',commentOnly:'0',demangle:'0',directives:'0',execute:'1',intel:'0',libraryCode:'0',trim:'1'),flagsViewOpen:'1',fontScale:14,fontUsePx:'0',j:1,lang:c%2B%2B,libs:!(),options:'-O2',selection:(endColumn:1,endLineNumber:1,positionColumn:1,positionLineNumber:1,selectionStartColumn:1,selectionStartLineNumber:1,startColumn:1,startLineNumber:1),source:1),l:'5',n:'0',o:'+x86-64+gcc+12.2+(Editor+%231)',t:'0')),k:50,l:'4',n:'0',o:'',s:0,t:'0')),l:'2',n:'0',o:'',t:'0')),version:4). I then replaced `-> f32` in the lambda with `__attribute__((always_inline)) -> f32` and got a lesser improvement:

```
/home/home/CLionProjects/iptsd/build/src/debug/iptsd-perf ./iptsdump.dat
[20:17:45.599] [info] Vendor:       045E
[20:17:45.599] [info] Product:      099F
[20:17:45.599] [info] Buffer Size:  7487
[20:17:45.599] [info] Metadata:
[20:17:45.599] [info] rows=44, columns=64
[20:17:45.599] [info] width=25978, height=17319
[20:17:45.599] [info] transform=[-412.3492,0,25978,0,-402.76746,17319]
[20:17:45.599] [info] unknown=1, [178,182,180,1,178,182,180,1,90,171,100,20,172,177,175,2]
[20:17:51.037] [info] Ran 2010 times
[20:17:51.037] [info] Total: 5427634μs
[20:17:51.037] [info] Mean: 2700.32μs
[20:17:51.037] [info] Standard Deviation: 933.69μs
[20:17:51.037] [info] Minimum: 146.740μs
[20:17:51.037] [info] Maximum: 4833.335μs
```

Disassembly of that shows 2 references to `sub_6870` to `hypotf@GOT`. The irrelevant one is `_ZN5iptsd5debug4perfL4mainEN3gsl4spanIPcLm18446744073709551615EEE` which seems to be `CLI::Option`'s internals. The other one is `_ZN5iptsd8contacts8advanced3alg4conv7kernels8gaussianIfLi5ELi5EEENS_9container6KernelIT_XT0_EXT1_EEES8_` which is in the advanced detection, but is far from a bottleneck as it barely shows as a vertical line in the perf graph. It is further confirmed by seeing the values after `vaddss vmovss vfmadd231ss vmovss vaddss vminss`. That's for the `dist == 0`, case, while the `dist = sqrt(2)`'s assembly looks more complicated (loading a float in a register, immediately saving it to the stack, calling a function, then loading the same value from the stack for no good reason) so it might not have been optimized properly. So inlining seems to be making a difference by removing that function call, but it fails to explain the other 10% improvement that the desugared template function gave us. Disassembly also reveals that the other lambdas below it such as `_ZZN5iptsd8contacts8advanced12BlobDetector7processERKNS_9container5ImageIfEEENKUliE5_clEi` aren't getting inlined either, possibly because of std::vector bounds checks.

I agree that compiler had trouble with the complexity. There are 43 calls to `get_cost`, of which there are 8 unique values for `index2_t d`. It may have run out of time during its optimization passes to realize that those 8 values simplify to 2 constant values in the end. On top of that there is `evaluate` adding to the layers of indirection the compiler needs to see through. The templates work here by forcing the compiler to consider what it needs to know.