@danielzgtg 2023-02-24
I changed `*` to `std::reference_wrapper`. For some reason that made it go to `2500us`, an increase of `100us` but still faster than how it was originally. It's strange because it's still the same vector and the `reference_wrapper` class doesn't do much.

I found `unchecked()` in `access.hpp`. I see it used in other parts of this advanced detector. In the assembly, I did see that there were calls related to std::vector. When that is used for the array accesses, we gain another 3% improvement and reach `2320us`:

```
/home/home/CLionProjects/iptsd/build/src/debug/iptsd-perf ./iptsdump.dat
[21:47:32.988] [info] Vendor:       045E
[21:47:32.988] [info] Product:      099F
[21:47:32.988] [info] Buffer Size:  7487
[21:47:32.988] [info] Metadata:
[21:47:32.988] [info] rows=44, columns=64
[21:47:32.988] [info] width=25978, height=17319
[21:47:32.988] [info] transform=[-412.3492,0,25978,0,-402.76746,17319]
[21:47:32.988] [info] unknown=1, [178,182,180,1,178,182,180,1,90,171,100,20,172,177,175,2]
[21:47:37.656] [info] Ran 2010 times
[21:47:37.656] [info] Total: 4657879μs
[21:47:37.656] [info] Mean: 2317.35μs
[21:47:37.656] [info] Standard Deviation: 789.74μs
[21:47:37.656] [info] Minimum: 147.491μs
[21:47:37.656] [info] Maximum: 3913.753μs
```